Disconnect BincodeDatabase from Cantine
Now that I can defer touching the database until rendering time,
I expect it will be a lot easier to juggle it separately when I
start dispatching tasks to different executors.

+++ crates/cantine/src/model.rs
@@ -107,7 +107,7 @@ pub struct SearchQuery {
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct SearchResult {
     pub items: Vec<RecipeCard>,
-    pub total_found: u32,
+    pub total_found: usize,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub agg: Option<FeaturesAggregationResult>,
